Update README.md
I tried to use the very same string shown in the doc for Brazilian Portuguese, which was "br", and it wasn't working right away. Debugging the code I saw that it must be "pt-BR" to be found in the "dates" at line 38. So I think that it would be nicer if the documentation could show the exact match for the option value (to all languages by the way) to make things clearer.