Clicking vertical scroll does not hide datepicker
I've been watching my users interact with the datepicker. I have it
attached to an input field with autoclose turned on. When the field is
near the bottom of the browser window and the user clicks on the date
field this causes the datepicker to appear partially off the bottom of
the browser window. They then use the vertical scrollbar to bring it
into view but it disappears upon clicking the scrollbar (at least it
does in IE). When it disappears the field retains focus so that clicking
on the date field again does not cause the datepicker to reappear. This
leaves them confused.

I suggest that clicking on the vertical scrollbar should not cause the
datepicker to hide itself.