Fixing Today + Active = black text on blue background
If todayHighlight == true and the active day in the calendar is today, the text remains black, despite the background being a dark blue (unreadable). I tracked it down to the !important declaration on the style definition for td.today. Removing !important doesn't seem to have any side effects in tested browsers: Chrome 25.0.1364.152 m and IE7-10.