Improve the visual responsiveness of borders
Improve the way that borders, for each picker (when using an horizontal
layout), behave to changes on the window size.

Use borders in every side of the pickers, as opposed to just right and
top, and simulate a collapse of the borders by using margin-right: -1px.

This change has the negative side effect of forcing the use of borders
in every situation, which doesn't look as clean as the precious version
when using a big window, but, in return, it allows a correct
visualization of the pickers in any situation (big windows, small
windows, inline datepicker inside another element with an area smaller
than the full width of the screen, etc).