Problem: Our translation team prefers the usage of [month number]月 for month selection as this is more commonly seen (one big example is the Windows date-time picker). This also allows better spacing for November and December which are 3 characters long in zh-cn and they run together.  Also the title text was built with string concatenation for month + year and does not leverage a localization format.
Solution: Switch monthsShort for zh-cn/zh-tw to use monthsShort from moment.js. Add Korean language support by porting from moment as well. In addition a titleFormat key has been added for the .datepicker-switch header of the date picker. This text used to be created with string concatenation and did not work well for Chinese and Korean. Chinese and Korean provide a titleFormat key override and this key defaults to "MM yyyy" for English and other languages.