Fixing month in title
Using the viewDate for displaying month in title. In commit 67a2482691d6c253064d44a47f83cde7b21cd6e3, this line was changed to introduce improved translation. The old version used the variable "month" which is based on viewDate. The new version created a new Date at the start of the month.

In my timezone, +2, the result of new Date(2015, 7) is "Sat Aug 01 2015 00:00:00 GMT+0200 (W. Europe Daylight Time)".  DPGlobal.formatDate then uses date.getUTCMonth(), which is now 6 (instead of 7). This again mean that we end up with "July 2015" in the title.