Add "assumeNearbyYear" option
Rebase / modification of #1556

If the user opts-in by setting options.assumeNearbyYear = true:

- When a user enters a 2 digit year, the code now does a better job of
guessing intent. If the 2 digit year is more than 10 years in the future,
it'll use the past century instead. So in the 2015, a value between 00-25
would assume the century is 2000. For 26-99, it'll assume the user meant
1900 (e.g. 1/31/10 = Jan 31, 2010, 1/31/26 = Jan 31, 1926). People don't
generally enter future dates too far in advance, but it's very common
to enter dates from 50+ years ago.