Fixed z-index calculation error on IE11 (#2008)
In IE11, the zIndex is returned as a number, which has parentsZIndex returned as an empty array.  This empty array has a max value applied, which is -Infinity, which causes the zIndexOffset to not be applied.

Chrome and other browsers accidentally get correct behavior since the zIndex property is returned a string, which makes !== 0 return true.

To unify the behavior, I added parseInt to the comparison, and set a default value of the parentsZIndex to 0, so that the max value is always at least 0.