Update bootstrap-datepicker.js
Trigger(s) always return the last date in dates array when we use setDate, setDates, etc., to set dates. There is no way to get the clicked date (among already selected dates). This change will return the clicked date if updateViewDate is true, and the last date in dates array if updateViewDate is false.