Align calendar to input element instead of group
When there is an input group with a prepend (like a field label) the calendar aligns to the prepend, with this small change it will align to the input field instead. I have not done extensive testing so it could be a bad idea, but it seems to work in every situation I throw at it.